+++ doc/changelog.rst
@@ -140,6 +140,10 @@ Breaking Changes in 4.0
   opposed to
   the previous syntax which was simply ``if collection:`` or ``if database:``.
   You must now explicitly compare with None.
+- Empty (None, {})  projections for find commands are passed on to the server
+  as-is rather than the previous behavior which substituted in a projection of
+  ``{"_id": 1}``. This means that an empty projection will now return the
+  entire document, not just the ``"_id"`` field.
 
 Notable improvements
 ....................